+++ bottombar.ios.d.ts
@@ -1,2 +1,2 @@
-export declare class NativescriptBottombar {
-}
+export * from './src/common';
+export * from './src/ios/bottombar';
