+++ src/ios/bottombar.d.ts
@@ -1,0 +1,2 @@
+export declare class NativescriptBottombar {
+}
