+++ src/ios/bottombar.ts
@@ -1,0 +1,3 @@
+export class NativescriptBottombar {
+    // TODO iOS 3.0.0 Implementation
+}
